Try chef-winrm-fs refactor
