+++ .rubocop_todo.yml
@@ -34,7 +34,7 @@ Metrics/PerceivedComplexity:
 Naming/HeredocDelimiterNaming:
   Exclude:
     - 'spec/matchers.rb'
-    - 'winrm-fs.gemspec'
+    - 'chef-winrm-fs.gemspec'
 
 # Offense count: 2
 # Configuration parameters: MaxUnannotatedPlaceholdersAllowed, IgnoredMethods.
@@ -47,7 +47,7 @@ Style/FormatStringToken:
 # Configuration parameters: AllowedReceivers.
 Style/HashEachMethods:
   Exclude:
-    - 'lib/winrm-fs/core/file_transporter.rb'
+    - 'lib/chef-winrm-fs/core/file_transporter.rb'
 
 # Offense count: 5
 # This cop supports unsafe auto-correction (--auto-correct-all).
@@ -56,21 +56,21 @@ Style/HashEachMethods:
 Style/NumericPredicate:
   Exclude:
     - 'spec/**/*'
-    - 'lib/winrm-fs/core/file_transporter.rb'
-    - 'lib/winrm-fs/file_manager.rb'
+    - 'lib/chef-winrm-fs/core/file_transporter.rb'
+    - 'lib/chef-winrm-fs/file_manager.rb'
 
 # Offense count: 1
 # Configuration parameters: AllowedMethods.
 # AllowedMethods: respond_to_missing?
 Style/OptionalBooleanParameter:
   Exclude:
-    - 'lib/winrm-fs/file_manager.rb'
+    - 'lib/chef-winrm-fs/file_manager.rb'
 
 # Offense count: 1
 # This cop supports safe auto-correction (--auto-correct).
 Style/RedundantBegin:
   Exclude:
-    - 'lib/winrm-fs/file_manager.rb'
+    - 'lib/chef-winrm-fs/file_manager.rb'
 
 # Offense count: 4
 # This cop supports unsafe auto-correction (--auto-correct-all).
