+++ lib/chef-winrm-fs/core/file_transporter.rb
