+++ lib/chef-winrm-fs/core/tmp_zip.rb
