+++ lib/chef-winrm-fs/exceptions.rb
