+++ lib/chef-winrm-fs/file_manager.rb
