+++ lib/chef-winrm-fs/scripts/check_files.ps1.erb
