+++ lib/chef-winrm-fs/scripts/checksum.ps1.erb
