+++ lib/chef-winrm-fs/scripts/create_dir.ps1.erb
