+++ lib/chef-winrm-fs/scripts/delete.ps1.erb
