+++ lib/chef-winrm-fs/scripts/download.ps1.erb
