+++ lib/chef-winrm-fs/scripts/exists.ps1.erb
