+++ lib/chef-winrm-fs/scripts/extract_files.ps1.erb
