+++ lib/chef-winrm-fs/scripts/scripts.rb
